Fix cache behavior when deleting the penultimate line
When deleting at <cursor> given the buffer content
"<cursor><newline>something", the cache would end up with to cache
lines containing "something" before this change. This happened because
cluffer handles such a deletion by destructively merging the content
of the second line into the first one, reporting that first line as
modified and dropping the second line. Before this change, the cache
didn't react to the implicit deletion of the second line.

This change adds a loop that deletes all trailing cache lines after
the cluffer update protocol has finished.