Draw thin, lowered cursor so that text does not jump around
The previous cursor implementation had the advantage of not
overlapping the characters of the buffer text. The disadvantage was
that cursor movement required the characters to move around as well.

This implementation draws a thin cursor beneath the characters of the
buffer text to avoid the problem of moving characters.