Store boundary of drawn region in CONTEXT, adapt drawing functions
The CONTEXT class now has slots {MIN,MAX}-{LINE,COLUMN} which describe
the region being drawn when the cache is rendered. The additional
slots {MIN,MAX}-COLUMN/CONTENT contain functions that return the
minimum and maximum column of the (non-whitespace) content of a given
line.

Drawing functions use this information to restrict all operations to
the region being drawn. As a consequence, the FIRST-LINE and LAST-LINE
parameters of several functions are no longer needed and have been
removed.

The new helper function MAP-LINES-IN-RANGE helps with drawing
multi-line output such that the caller only has to draw individual
lines with suitable restrictions already applied.

The functions DRAW-UNDERLINE-MARKER and DRAW-MULTIPLE-LINE-RECTANGLE
use MAP-LINES-IN-RANGE and therefore now handle multi-line output as
well as restriction to the drawn region correctly.